# Speedup vm start by making vm_passwd saving much faster