--- a/systemvm/patches/debian/config/opt/cloud/bin/update_config.py
+++ b/systemvm/patches/debian/config/opt/cloud/bin/update_config.py
@@ -26,6 +26,8 @@ import os.path
 import configure
 import glob
 import json
+from cs.CsVmPassword import *
+
 
 OCCURRENCES = 1
 
@@ -48,13 +50,27 @@ def finish_config():
     sys.exit(returncode)
 
 
-def process_file():
+def process(do_merge=True):
     print "[INFO] Processing JSON file %s" % sys.argv[1]
     qf = QueueFile()
     qf.setFile(sys.argv[1])
+    qf.do_merge = do_merge
     qf.load(None)
+    return qf
+
+
+def process_file():
+    print "[INFO] process_file"
+    qf = process()
     # Converge
     finish_config()
+
+
+def process_vmpasswd():
+    print "[INFO] process_vmpassword"
+    qf = process(False)
+    print "[INFO] Sending password to password server"
+    CsPassword(qf.getData())
 
 
 def is_guestnet_configured(guestnet_dict, keys):
@@ -133,6 +149,10 @@ if sys.argv[1] and sys.argv[1].count("guest_network.json") == OCCURRENCES:
     else:
         print "[INFO] update_config.py :: No GuestNetwork configured yet. Configuring first one now."
         process_file()
+# Bypass saving passwords and running full config/convergence, just feed passwd to passwd server and stop
+elif sys.argv[1].startswith("vm_password.json"):
+    print "[INFO] update_config.py :: Processing incoming vm_passwd file => %s" % sys.argv[1]
+    process_vmpasswd()
 else:
     print "[INFO] update_config.py :: Processing incoming file => %s" % sys.argv[1]
     process_file()