… systemvm/patches/debian/config/opt/cloud/bin/cs_vmp.py
… systemvm/patches/debian/config/opt/cloud/bin/merge.py.rej