This speeds up preparing the router with the vm_password at vm start. The more VMs in a VPC, the bigger the difference this PR makes.

Changes:
- do not keep passwords in databag (/etc/cloudstack/vmpasswd.json). It has no value because the passwd_server already saves them
- process only the password we get in (vm_password.json) from mgt server (instead of all passwords all the time)
- lookup the correct passwd server instead of adding the passwd to all of them (one per tier)

Example:
- 4 tiers and 199 VMs running
- Start vm 200 would cause new passwd from vm_password.json (1) to be merged with /etc/cloudstack/vmpasswd.json (199)
- A curl command was exected foreach password (200) foreach tier (4) resulting in 800 calls
- In fact, since passwds are never cleaned it could very well be even more as the ip address was the key in the json file so until the ip address was reused the original password would remain and be sent to passwd server every time another vm starts.
- This took ~40 seconds of the time to start a VM (per router)

With the changes in this PR we just figure out the right tier and only process the new password resulting in a single curl call.

Setting passwd now takes 0.03sec regardless of the number of VMs or tiers.

```
2016-05-07 18:26:51,005  merge.py load:256 Continuing with the processing of file '/var/cache/cloud/vm_password.json.c0d4f015-12e3-49e5-882f-2aae3301694d'
2016-05-07 18:26:51,006  merge.py __moveFile:298 Processed file written to /var/cache/cloud/processed/vm_password.json.c0d4f015-12e3-49e5-882f-2aae3301694d.gz
2016-05-07 18:26:51,006  merge.py load:269 Not merging because do_merge is False
2016-05-07 18:26:51,006  CsVmPassword.py __update:30 Got VM '10.4.1.53' and password 'pnwJS5Ad3gsz'
2016-05-07 18:26:51,006  CsHelper.py execute:184 Executing: ip addr show | grep inet | grep -v secondary | awk '{print $2}'
2016-05-07 18:26:51,016  CsVmPassword.py __update:33 Found these CIDRs: ['127.0.0.1/8', '169.254.3.39/16', '85.222.237.187/25', '10.4.1.18/24', '10.4.2.243/24', '10.4.3.127/24', '10.4.4.109/
24']
2016-05-07 18:26:51,016  CsVmPassword.py __update:35 Processing CIDR '127.0.0.1/8'
2016-05-07 18:26:51,016  CsVmPassword.py __update:35 Processing CIDR '169.254.3.39/16'
2016-05-07 18:26:51,016  CsVmPassword.py __update:35 Processing CIDR '85.222.237.187/25'
2016-05-07 18:26:51,016  CsVmPassword.py __update:35 Processing CIDR '10.4.1.18/24'
2016-05-07 18:26:51,016  CsVmPassword.py __update:38 Cidr 10.4.1.18/24 matches vm ip address 10.4.1.53 so adding passwd to passwd server at 10.4.1.18
2016-05-07 18:26:51,017  CsHelper.py execute:184 Executing: ps aux
2016-05-07 18:26:51,029  CsProcess.py find_pid:50 CsProcess:: Searching for process ==> ['/opt/cloud/bin/passwd_server_ip.py', '10.4.1.18'] and found PIDs ==> ['25500']
2016-05-07 18:26:51,030  CsHelper.py execute:184 Executing: curl --header "DomU_Request: save_password" "http://10.4.1.18:8080/" -F "ip=10.4.1.53" -F "password=pnwJS5Ad3gsz" -F "token=2b23d8
18236f1f87da9905345cd1c4ed" --interface 127.0.0.1 >/dev/null 2>/dev/null &
2016-05-07 18:26:51,035  CsVmPassword.py __update:44 Update password server result ==> []
2016-05-07 18:26:51,035  CsVmPassword.py __update:35 Processing CIDR '10.4.2.243/24'
2016-05-07 18:26:51,035  CsVmPassword.py __update:35 Processing CIDR '10.4.3.127/24'
2016-05-07 18:26:51,035  CsVmPassword.py __update:35 Processing CIDR '10.4.4.109/24'
```

Big difference in the nr of curl commands. Now, just one:

```
2016-05-07 18:26:51,030  CsHelper.py execute:184 Executing: curl --header "DomU_Request: save_password" "http://10.4.1.18:8080/" -F "ip=10.4.1.53" -F "password=pnwJS5Ad3gsz" -F "token=2b23d818236f1f87da9905345cd1c4ed" --interface 127.0.0.1 >/dev/null 2>/dev/null &
```

Before on a 4-tier, 200VM VPC:
```
2016-05-06 20:15:51,407  CsHelper.py execute:184 Executing: curl --header "DomU_Request: save_password" "http://10.4.1.18:8080/" -F "ip=10.4.1.89" -F "password=bla" -F "token=2b23d818236f1f87da9905345cd1c4ed" --interface 127.0.0.1 >/dev/null 2>/dev/null &
<cut>
2016-05-06 20:16:28,600  CsHelper.py execute:184 Executing: curl --header "DomU_Request: save_password" "http://10.4.4.109:8080/" -F "ip=10.4.1.79" -F "password=bla" -F "token=2b23d818236f1f87da9905345cd1c4ed" --interface 127.0.0.1 >/dev/null 2>/dev/null &
```

![image](https://cloud.githubusercontent.com/assets/1630096/15094309/89720608-149f-11e6-944e-82ae91545c7a.png)

```
cat curllog | wc -l
804
```

As you can see, it took ~40 seconds and more VMs or more tiers would make it worse and worse..

Now, it takes 0.03sec, so ~100 times faster!

In the previous perfect scenario (new VPC, one vm, one tier) this would also result in one curl call. Now it always uses one call regardless of the nr of VMs.