@remibergsma 2016-05-07
Passwd server tests:
```
test_01_vpc_password_service_single_vpc (integration.smoke.test_password_server.TestPasswordService) ... === TestName: test_01_vpc_password_service_single_vpc | Status : SUCCESS ===
ok
test_02_vpc_password_service_redundant_vpc (integration.smoke.test_password_server.TestPasswordService) ... === TestName: test_02_vpc_password_service_redundant_vpc | Status : SUCCESS ===
ok
```