@remibergsma 2016-05-07
Integration tests failed due to unrelated deploy issue: https://beta-jenkins.mcc.schubergphilis.com/job/cosmic/job/0300-setup-infrastructure-for-integration-tests/142/console