@sspans 2016-05-09
Why is this passed on via curl? Should be trivial to do in native python.