@remibergsma 2016-05-09
Build 70 has 7 failures, build 71 (same branch) has 4. Some of them are failing for a couple of builds already as I see the same ones in build 67.